migrate view tracking to websocket and filter zombie sessions
