+++ server/src/services/websocket.rs
@@ -49,10 +49,51 @@ pub async fn ws_handler(
 
 async fn handle_socket(socket: WebSocket, state: AppState, session_id: String, user_id: Option<i64>) {
 
-    {
+    // Track if this is a first-time connection for view counting
+    let was_previously_connected = {
         let mut map = state.lock_sessions();
         if let Some(ctx) = map.get_mut(&session_id) {
+            let was_connected = ctx.is_ws_connected;
             ctx.is_ws_connected = true;
+            was_connected
+        } else {
+            false
+        }
+    };
+
+    // Count view ONLY on first WebSocket connection (not on HTTP GET)
+    if !was_previously_connected {
+        let session_ctx = {
+            let map = state.lock_sessions();
+            map.get(&session_id).cloned()
+        };
+
+        // Only count views for viewer sessions (not builders)
+        if let Some(ctx) = session_ctx {
+            if let (Some(owner_id), Some(slug)) = (ctx.project_owner_id, &ctx.project_slug) {
+                let db_clone = state.db.clone();
+                let slug_clone = slug.clone();
+                tokio::spawn(async move {
+                    // Increment view counter
+                    let _ = sqlx::query("UPDATE projects SET view_count = view_count + 1 WHERE owner_id = $1 AND LOWER(slug) = LOWER($2)")
+                        .bind(owner_id)
+                        .bind(&slug_clone)
+                        .execute(&db_clone)
+                        .await;
+                    
+                    // Log View event for analytics
+                    if let Ok(Some(project_id)) = sqlx::query_scalar::<_, i64>(
+                        "SELECT id FROM projects WHERE owner_id = $1 AND LOWER(slug) = LOWER($2)"
+                    )
+                    .bind(owner_id)
+                    .bind(&slug_clone)
+                    .fetch_optional(&db_clone)
+                    .await
+                    {
+                        log_analytics_event(&db_clone, project_id, AnalyticsEventType::View, None, None).await;
+                    }
+                });
+            }
         }
     }
 
@@ -427,9 +468,9 @@ async fn attach_to_container(
         if should_delete {
             println!("Cleaning up session: {}", session_id);
             
-            // Log session end event for viewer sessions (not builders)
+            // Log session end event only for viewer sessions that actually connected
             if let Some(ctx) = session_ctx {
-                if ctx.project_slug.is_some() && ctx.project_owner_id.is_some() {
+                if ctx.is_ws_connected && ctx.project_slug.is_some() && ctx.project_owner_id.is_some() {
                     let duration = ctx.created_at.elapsed().as_secs() as i64;
                     
                     // Lookup project_id from slug and owner
